Add intro link image as slice
