+++ apps/web/components/Organization/Slice/IntroLinkImageComponent/IntroLinkImageComponent.css.ts
