+++ apps/web/components/Organization/Slice/IntroLinkImageSlice/IntroLinkImageSlice.tsx
@@ -1,0 +1,11 @@
+import type { IntroLinkImage } from '@island.is/web/graphql/schema'
+
+import { IntroLinkImageComponent } from '../IntroLinkImageComponent/IntroLinkImageComponent'
+
+interface IntroLinkImageSliceProps {
+  slice: IntroLinkImage
+}
+
+export const IntroLinkImageSlice = ({ slice }: IntroLinkImageSliceProps) => {
+  return <IntroLinkImageComponent id={slice.id} item={slice} />
+}
