+++ apps/web/components/real.ts
@@ -101,3 +101,4 @@ export * from './FeaturedLinks/FeaturedLinks'
 export * from './Organization/Slice/LatestNewsSlice/DigitalIcelandLatestNewsSlice'
 export * from './DigitalIcelandNewsCard/DigitalIcelandNewsCard'
 export * from './AddToCalendarButton/AddToCalendarButton'
+export * from './Organization/Slice/IntroLinkImageSlice/IntroLinkImageSlice'
